Modified the handling of patterns generated with the RF command to allow both colored and uncolored variations.
Modified the manner in which the current pen is specified for
uncolored patterns generated via the RF command. If pen 1 is desired,
the opposite of the current pen is specified, to allow for the
possibility that the pattern may not exist.

Changed the handling of fill adjust to reflect the replacement of the
grid_adjust state parameter with pp_mode (note that the actual
implementation of pixel placement for GL objects is still broken).

git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@369 06663e23-700e-0410-b217-a244a6096597