Modified to support color-mapping as part of a rendering method, and to support monochrome mode.
As part of this change, the never-used device field of the rend_info_t
structure is replaced with a color mapping parameter, which is used
with a color-mapping forwarding device that is always present.

To support monochrome mode, the nature of the remapping array has been
change somewhat, and the render_method field of the pcl_ht_t structure
modified to be the mapped rather than raw rendering method. Fields for
print mode (monochrome or color) and raw rendering method have been
added to the PCL state.

git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@388 06663e23-700e-0410-b217-a244a6096597