Changed from a single color mapping device to multiple color mapping devices, one for each mapping type. This is needed because the rendering method used by the foreground may be different from that used by the palette (which, in turn, is used by a raster), so it is necessary to have two different color mapping device present at the same time.
Added support for table-dither (mask, non-monotonic) halftones to be
used with the (not fixed) built-in rendering methods.

Added code to allow device-specific halftones and re-mapping arrays
encoded as gs_param dictionaries to be read from the device on start-up.

Added support for table dither halftones. Also provided a structure
descriptor for pcl_ht_builtin_dither_t structures, as these may now need
to be allocated when read from a device.

git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@403 06663e23-700e-0410-b217-a244a6096597