Turned off the hack that forced all TrueType fonts "symbolic" This fixes invisible text in acrobat 5.0.
Hacked in a test for character type in pdf_text_process it inspects
the penum.text character data to determine if it is packed 8 bit
or 8 bit unpacked into shorts or 16bit values.
This is fragile and should be replaced with get_next_char routines that
will function given the availiable state in pdfwrite.

git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@1392 06663e23-700e-0410-b217-a244a6096597