Added conditional compile code based on linkage with Postscript interpreter. PSI_INCLUDED Ideally this conditional would not be needed.
It currently adds the ps interpreter in plimpl.c, this would be easy to
change to a choice of plimpl.c files in the applicaton directory.

It currently incorrectly sets the device based on postscript.
Experimental and non-working.

It also switches the memory allocator to the ps allocator in gsmemory.c

Experimental conditionals in plalloc.c allow the removal of memory globals.

git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@1457 06663e23-700e-0410-b217-a244a6096597