Added Initial Matrix with support for scaling and translation. Added translation and scaling device switches:
switch------ = default : range
-dViewScaleX = 1.0 : 1.0 to 1000+
-dViewScaleY = 1.0
-dViewTransX = 0.0 : 0.0 to (width * ViewScaleX)
-dViewTransY = 0.0

scalefactors less than one will leave a white border as there is more
buffer space than data.

Translations need to have the scalefactor multplied in.  This allows all
pixel addresses to be translated to the origin.  If this were done
automatically by the initial matrix then only address that are muliples
of the HWResolution could be translated to 0,0.

git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@1493 06663e23-700e-0410-b217-a244a6096597