Added feature ENABLE_AUTO_GRAY_RENDER_METHODS, too match hp clj4500, and 4550
When set to true it automatically maps render method to a corresponding
gray render algorithm if the entire palette only contains gray entries.

When set to false the input render method is observed.

PCL lacks a gray color space.  This auto gray mapping has two benefits,
on a multi-pass color printer it is faster to only use K, and by not using
CMY in a gray palette it is much easier to get a gray value that matches
the white point of the paper (no visible color in the gray).

git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@1502 06663e23-700e-0410-b217-a244a6096597