Navigtor window that drives a Gview window.
It is slow since two executions of pcl6 occur, one at low res and one at high
res.

Translations an zooms on a page in the Nav window affect the high res window,
with a red box showing you where you are.

The high res window is not a full frame just a view port so very high
rendering resolutions are possible on a low memory machine.

Same command set as Gview.

git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@1523 06663e23-700e-0410-b217-a244a6096597