Added memory cache of pages.  Default of 10 pages are cached in a circular queue.  Cache entries are replace in a last in first replaced fashion. Enries are invalidated if the parameters for the page requested have changed, zooming for instance.
Note: It maybe better to store pages as images on disk, this would allow
more images to be cached.

Added look ahead.  When idle the next page is loaded into cache in the
background, the intent is that if the document is being viewed for a few
seconds this time can be used to load the next page.

Currently the generation of page counts in the background prevents
look ahead until it is complete.  It doen't prevent navigation of the document.

Moved the responsibility of page number to fetch, and total number of pages in
the document up away from Gpickle.  This was needed since one instance of
Gpickle is used to get pages and another is used to get page counts.
A one to one relationship between the number of threads ( GpickleThread ) and
simultanous render processes ( Gpickle ) must be maintained.

git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@1570 06663e23-700e-0410-b217-a244a6096597