XL format 0 fonts class 1 fonts can be downloaded with 3 segments instead of 5 to conserve memory.  Escapement and [left|top]sidebearing can be embedded in the XL character header and the hhea and hmtx segments omitted.  Error checking routines improperly rejected these fonts.
In some cases the UFST font scaler was returning a bitmap when a
character outline was required.

Recently we have seen XL format 0 class 1 and 2 fonts that use the
character code as a key instead of the glyph id embedded in the
downloaded font data.  It is possible that the character code is also
used for format 15 and XL format 0 class 0.  Pending further
investigation we use the character code as the key for format 0 class
1 and 2 and the glyph id embedded in the font (what one might expect)
for the other font types.

git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@1605 06663e23-700e-0410-b217-a244a6096597