Changed from single string command line to array of strings, to support pjl which embeds spaces in a single argument which confused the java default parsing of command line arguments.
Add RTL options for playing with plotsize overrides.

Changed the search for pcl6 exectable to first look relative then look in the current
directory and as a last resort look for pcl6.exe in the path.  Note the exe is for
windows and unix, sorry but I choose not to search the users path.

git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@1646 06663e23-700e-0410-b217-a244a6096597