gs_exit() cleanup.
gs_exit() is renamed gs_to_exit() since it is an about to terminate cleanup
function, it always returns.

Both gs_to_exit() and gs_abort are client provided functions.  The client
can have its own special cleanup on exit, and the client can decide what it
wants to do in the abort case: exit, setjump/longjump, ...

gp_do_exit() is the way for the client to call a system independent exit()
this allows for different exit codes on unix and vms for instance.
Only the client can call this, if so desired from their gs_abort().

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@2638 a1074d23-0009-0410-80fe-cf8c14f379e6