Fix for [ 652147 ] overprintmode not boolean.  Adobe TN5145 defines the parameters to/from setoverprintmode/currentoverprintmode as being boolean.  This change modifies the implementation to match the TN.
DETAILS:

Previously the operators setoverprintmode and currentoverprintmode had
been defined to match the requirements of PDF which uses an integer for
the parameter.  However Adobe TN5145 defines the postscript implementation
as using a boolean value.  This change modifies the definition of these
operators in lib/gs_ll3.ps to convert the parameter type.  The internal
operators .setoverprintmode/.curentoverprintmode still use the PDF integer
parameter.  The postscript code in lib/gs_ll3.ps converts the parameter
type.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@3426 a1074d23-0009-0410-80fe-cf8c14f379e6