Reduce state.width delta radius calculation for shadings that was made too large by the previous commit.
DETAILS:
Previous commit mistakenly included the state.dr value as both the X and
Y component in the distance transform that produced the dev_dr vector.
The hypot is actually all that is required to handle getting the correct
scaled radius distance independent of the CTM rotation.

The use of state.dr for both X and Y preceded the use of hypot on dev_dr
and wasn't taken out when the hypot function use was added.

This should eliminate most, if not all regressions caused by the 12/24
1.18 patch to this module.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@3457 a1074d23-0009-0410-80fe-cf8c14f379e6