Within the display device, correctly allocate the memory device with the correct structure descriptor.
DETAILS:
The display device was allocating a memory device by taking a
memory device prototype from gdev_mem_device_for_bits and then
using gs_copydevice on it.  This allocates the wrong size
because the prototype does not have a structure descriptor.  The
correct way is to allocate it with gs_alloc_struct with
st_device_memory as an argument.  This bug was found when using
the display device in GhostPCL, which was complaining on
shutdown about the unfreed structure descriptors left behind by
gs_copydevice.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@3535 a1074d23-0009-0410-80fe-cf8c14f379e6