Fix for [ 664473 ] 64bit color indexes break X driver. This fix corrects some problems found with the X window driver and 64 bit gx_color_indexes found by customer 1110.
DETAILS:

The problems were found on Sparc workstations running the
Solaris 8.0.  The problem has not been reported on other
systems.  However the fix is simple (and contributed by
customer 1110) and should work on all systems.  The fix
is simply to use "long" for values passed to X windows
system routines.  This is done via a combination of casts
or local variables depending upon the situation.  The use
of "long" for gx_color_index was the default prior to the
8.0 release.  The X windows device does not need the extra
size for color values.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@3541 a1074d23-0009-0410-80fe-cf8c14f379e6