Treat files starting with %Pxxx as PostScript, not PDF.  Process empty files without error.  Fixes SourceForge bug #663150, Some PS files half-recognized as PDF files
DETAILS:
Changes for the 8.00 release to allow reading of PDF files from
stdin caused files starting with %Pxxx to be half recognised as
PDF files.  This patch simplifies the PDF recognition code and
makes it treat only files starting with "%PDF-" as PDF.  Testing
of this code found that zero length input files caused a
PostScript error.  This was an existing error which is fixed by
the gs_epsf.ps patch and last line of the pdf_main.ps patch.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@3545 a1074d23-0009-0410-80fe-cf8c14f379e6