Fix regression of 119-01.ps caused by previous patch while retaining the fix for #667658 when CR and/or LF comes between the '~' and the '>'.
DETAILS:

The logic was also fixed so that if the end of buffer is reached while
searching for the '>' character (past CR and/or LF) no error would be
issued unless we have the 'last' condition indicating that no more data
is available.

Thanks to Igor M. for spotting the above potential problem during code
review.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@3576 a1074d23-0009-0410-80fe-cf8c14f379e6