Fix : Eliminate .forceput from gs_resmp.ps .
DETAILS:

This patch should not change any behavior.

A mapped category used an array to store a set of procedures.
Sometimes the set grows, and the old implementation
used .forceput to replace the array with a bigger one
in the category dictionary, which is read-only.

The new code represents the set as a dictionary,
so that .forceput is not longer needed.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@3578 a1074d23-0009-0410-80fe-cf8c14f379e6