Fix for 673108.  The test file has a stream which contains an invalid stream length.  However Acrobat handles the file.
Details:
The invalid stream length was bad enough that we were positioning beyond
the end of the file when we searched for the endstream.  A check was
added for this case and if so then we ignore repositioning the file
before searching for the endstream.  Note:  There is already logic in
the code which emits a warning message if the endstream is not found.
A second part of this fix was to clean up the stack if the endstream
is not found.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@3584 a1074d23-0009-0410-80fe-cf8c14f379e6