Fix (pdfwrite) : -dSubsetFonts=false did not work with incremental fonts.
DETAILS :

pdfwrite created a "complete" copy of a font at the first appearance of the font.
With incremental fonts it missed incremental glyphs with -dSubsetFonts=false.
The patch checks whether a glyph is added incrementally and
replaces the "complete" font copy with a font subset.
Bug #686735 "-dSubsetFonts=false produces a bad PDF".

The core of the patch is the change to gdevpdte.c .
Other changes just pass a necessary data through internal interfaces.

This patch should not change behavior with -dSubsetFonts=false,
which is used by default for the regression testing.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@3668 a1074d23-0009-0410-80fe-cf8c14f379e6