Try to make autogen.sh smarter about automake version mismatch.
This is really silly, but there's a serious problem with incompatibility
between the various versions of automake, and while there is a mechanism
for requesting a particular version in the Makefile.am, most
installations don't make use of this information, and you're supposed to
know which automake to call. Since this obviates the point of autogen,
we attempt to match a requested version in AUTOMAKE_HEADERS to likely
executables. Probably not very robust.


git-svn-id: http://svn.ghostscript.com/jbig2dec/trunk@213 ded80894-8fb9-0310-811b-c03f3676ab4d