Fix for #686755 sgirgb device.  The sgirgb device would produce compiler errors if included in version 8.00 or later.
DETAILS:

The sgirgb device had not been modified to match the changes made in the
device API with version 8.00.  These were fixed.  In addition a couple of
compiler warnings were removed.  AN unsed variable (sptr) was also
removed.

Note:  I am unable to test this fix since I do not have an sgi system.
The inclusion of the sgirgb device also creates link errors in a WinXP
build due to undefined/missing bzero.  Since this was not changed, I
am assuming that this is not a problem on an sgi system.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@3696 a1074d23-0009-0410-80fe-cf8c14f379e6