Replace the win32-specific code for seeking with a generic version. Partial fix for bug #686739.
DETAILS:

The code was making an assumption about the internals of the opaque
fpos_t type. This is a 64 bit integer on win32, but a struct on
linux and macos, so the compile of course fails.

This commit temporarily breaks diff.bmp output on win32, which allegedly
cannot fseek() beyond the written end of the file so the row-reversal
logic will fail. This will be corrected in another revision.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@3713 a1074d23-0009-0410-80fe-cf8c14f379e6