Fix (pdfwrite) : When advance width has non-zero y component, the width was set to zero. Bug 686764 "Overlapping characters".
DETAILS :

The test document embeds a font, in which some characters
have a small non-zero y component of the advance width.
It looks as a buggy font. The pdfwrite writes zero to Widths array
(that's correct, because it is only useful with zero y component),
and did not generate glyph positioning commands (the bug).
This patch makes pdfwrite to be tolerant to such buggy fonts
with generating glyph position for subsequent glyph.
It does not repair the font.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@3725 a1074d23-0009-0410-80fe-cf8c14f379e6