Fix (file_name_combine) : Check .tempfile permissions.
DETAILS :

The temporary directory is allowed for creating temporary files
without specifying a special permission. However when the
prefix argument of .tempfile is an absolute path,
the permission is checked for it.

Minor change :

On Windows gp_open_scratch_file now calls gp_gettmpdir,
making the logic closer to one on Unix.
The call to GetTempPath is left for backward compatibility
(when the environment specify TMP but no TEMP).

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@3881 a1074d23-0009-0410-80fe-cf8c14f379e6