Fix : margin_set::y was not initialized starting a fill.
DETAILS :

It could indereministically execute some code in start_margin_set,
but the code appears idle in this case because margin_set::margin_list == NULL.
Therefore it could not cause an indeterministic result.
Therefore this patch is an equivalent change, which is useful
to simplify the job for code validation tools.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@3897 a1074d23-0009-0410-80fe-cf8c14f379e6