Log message: For the pngalpha device, set the default background color to white in the bKGD chunk.  Add an option -dBackgroundColor=16#RRGGBB to explicitly set the background color.  This is needed by a common web browser that doesn't correctly support PNG with alpha transparency.
DETAILS:
Programs that don't fully support alpha transparency let
libpng convert an RGBA image to RGB.  The background color
used is either provided by the program, or in the bKGD
chunk, or it falls back to gray.  MSIE was defaulting
to a gray background which is not commonly used on web
pages.  By writing a white bKGD chunk, PNG files will display
correctly against a white background in MSIE.
If a solid color for a web page background is used
with <body bgcolor="#CCCC00">, then PNG images appearing
on that page would need to be created with
-dBackgroundColor=16#CCCC00.
For textured background there is no hope.
Microsoft needs to fix Internet Explorer.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@3910 a1074d23-0009-0410-80fe-cf8c14f379e6