The TrueType font loading was quite a bit slower than it should be. This optimized font loader is up to ten times faster. Fixes bug #686851.
DETAILS:

Analysis of the time spent during TT font loading led to two problem areas.

The first was that the cmap array was built inefficiently when their was
a large glyph range, that had large holes was built with a loop that
called a PostScript function 'putglyph' to fill unused areas with .notdef
(glyph ID 0). The revision pre-fills the 'glyphs' array with .notdef from
a static array pre-filled with 1024 zeroes.

The second problem was the code that built the encoding using the 'post'
table. This code had a 'repeat' loop to find the name from the start of
the glyphNames section of the 'post' table for each named glyph in the
encoding. By creating a temporary cross reference, only one pass through
the names is made.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@3936 a1074d23-0009-0410-80fe-cf8c14f379e6