Fix (Type 1 hinter) : -dAlignToPixels=0 was ignored with -dNOCACHE.
DETAILS :

Now we pass the oversampling scale and the number of subpixels
to the Type 1 hinter in separate parameters. The old interface
was wrong when alpha buffer is installed with -dNOCACHE.

The patch cause difference in rendering with -dTextAlphaBits=2,4.
These cases are not tested by the regression test tool.

Minor changes :

- fixed C-style in gxhintn.c;
- removed "const gs_log2_scale_point no_scale = {0, 0}" to simplify calls
  to gs_type1_interp_init.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@4044 a1074d23-0009-0410-80fe-cf8c14f379e6