Fix (pdfwrite) : Enlarge font resource arrays on demand.
DETAILS :

The patch fixes CID fonts that provide a lesser CIDCount than
CIDs used in a document. Rather PS requires to print CID=0,
we need to provide a bigger CIDCount since we don't
re-encode the text. The text should look fine if the
viewer application substitutes the font.

Fixes the pdfwrite regression with 401-01.ps, which caused a segfault on Linux.
Thanks to Ray Johnston for the fault analyzis.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@4050 a1074d23-0009-0410-80fe-cf8c14f379e6