Add an initial outermost save unless -dNOOUTERSAVE option is specified. Fixes Bug #686972 and others that were closed with the instructions to use -c "save pop" prior to the file.
DETAILS:
The initial save is needed for several ProcSets from Adobe applications
and those of other vendors that expect objects in global VM to be retained
after a restore that would otherwise remove objects in global VM defined
after the corresponding save.

This kept recurring because several ProcSets from Adobe, Quark and others
expect to run under a job server. The -dNOOUTERSAVE is needed for backward
compatibility and for those that want to run under a job server and want
global VM to be restored between jobs (as expected under a job server).

Note that the save object is discarded so that it cannot be used to restore
to the state prior to .setsafe and thus circumvent security when -dNOSAFER
is used and .setsafe or .locksafe is subsequently used to go to SAFE mode.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@4092 a1074d23-0009-0410-80fe-cf8c14f379e6