Fix (pdfwrite) : Providing an access to the original font matrix via a special font dictionary key /.OrigFont .
DETAILS :

The new key is implementation specific and should not affect regular documents.
It works for the case when a font is copied like this :

	/x findfont dup length dict copy
	/y exch definefont

- it points from the copy to the original font.

This change fixes Ghostscript Bug 686970 "Overlapping characters"
and causes a progression with 455690.pdf with pdfwrite.

A long ago Peter L. Deutsch and me discussed the heuristic in
pdf_font_orig_matrix/font_orig_scale. I was aware of its limitations,
but I followed Peter's authority. Now we get a real case that it can't handle,
and therefore I take the responsibility for the different approach.

-igor.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@4103 a1074d23-0009-0410-80fe-cf8c14f379e6