Fixes two pointers to freed objects that were left dangling in the PDF 1.4 transparency logic. Fixes bug #686994.
DETAILS:

With some test files, saved graphics states had held on to pointers to
the device filter stack that had already been freed by a
gs_pop_device_filter() call. These dangling pointers can cause garbage
collection corruption. This patch introduces reference counting
discipline to these pointers so that the dfilter_stack objects are not
actually freed until the last reference is dropped.

A similar but simpler argument applies to the pdf14_ctx object in the
pdf14_device. This patch simply sets the pointer to NULL when freeing
the context.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@4135 a1074d23-0009-0410-80fe-cf8c14f379e6