Implementing a serialization for color spaces and functions.
DETAILS :

This stuff will be used by high level devices for comparing color spaces.
A serialization will. be used because we need to compare color spaces,
which's life times don't intersect.

This patch only defines new functions but never calls them.
Regressions should not appear.
A draft implementation for the functions is provided.
A debugging of them will be a separate issue.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@4197 a1074d23-0009-0410-80fe-cf8c14f379e6