pdfwrite : A minor code cleanup.
DETAILS :

'sclose' is inserted for a conformity, rather it has no effect in this case.

The "fixme" comment is removed because it appears wrong.
If the color space is new, it needs a complete serialization
for storing the data to PDF resource. We can't delay the serialization
because the life time of the original color space may exceed before
the next comparison happens. Contrary, if the color space is not new,
it also needs a complete serialization for a full comparison,
to know whether it is same color space.
Thus, a complete serialization is required in any case.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@4201 a1074d23-0009-0410-80fe-cf8c14f379e6