gdevdrop.c: Fixes raster operation with bit tagging device.  The rop code attempted to handle CMYK devices by assuming a default transformation between DeviceCMYK and RGB.  This doesn't work for the bit tagging device or any other device for that matter.
gdevbit.c: Set bandingAlways for the bit tagging device.  Banding
works reasonably well.  It misclassified a few objects in the pcl and
pxl fts.  But I believe these problems can be addressed quickly, so we
set banding on in the development code.  Also in this file the bit
twiddling to convert color to rgb was slightly off for the exta tag
byte and we now enable bit tagging in the device's put_parms() call.
The previous initialization in get_params did not work correctly for
XL's initial fillpage call.

dotags.sh: modified to take multiple files on the command line.

git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@1975 06663e23-700e-0410-b217-a244a6096597