Changed memory limit checking to decouple garbage collection activity from a non gc system.
When the garbage collector isn't installed the base class
(gs_memory_t) is responsible for managing an artificial limit
on allocations.

When gc is enabled the gs_ref_memory->limit is used as a high
water mark to signal a gc cycle.  The value is dynamic and not
suitable for a maximum memory limit.

git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@1976 06663e23-700e-0410-b217-a244a6096597