Propagating return code from gs_main_finit to GSDLL API. THIS IS INCOMPATIBLE CHANGE : the API had changed a prototype.
A close_device method can terminate with an error,
which was not delivered to clients.
It is especially important with pdfwrite,
which performs a big work in close_device.

Now clients can know about the error checking the return code
of the gsapi_exit function. See examples in doc/API.htm .

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@4249 a1074d23-0009-0410-80fe-cf8c14f379e6