Change the name of NOSUBCOLORSPACES to NOSUBDEVICECOLORS.
DETAILS:

In a code review for the NOSUBCOLORSPACES switch, Igor suggested changing
the name to KEEPDEVICECOLORS.  I still prefer NOSUB instead of KEEP.
Basically I prefer that switches which stop an action begin with NO.
However changing COLORSPACES to DEVICECOLORS is a definite improvement
since COLORSPACES is overly general and thus inaccurate.  The only
color spaces which are substituted are the device colors.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@4274 a1074d23-0009-0410-80fe-cf8c14f379e6