Fix a bug with text region rendering in the Patent.pdf example file from the adobe's encoder.
The problem was that we weren't handling the export/import of symbols 
from dictionaries referred to by the symbol dictionaries referred to by 
the text region segment.

This patch is just a quick fix: we recurse when building the list of 
referred to symbol dictionaries at text region decode time, which 
allows us to get at all the symbols. This works because the adobe 
encoder (at least in this file) always exports all the symbols, imported 
and encoded, in each dictionary. If it did not, we would have offset 
errors in the symbol lookup.

A proper fix requires decoding the exported symbol flags and building an 
export version of SBSYMS to store in the segment result that includes 
the symbols imported from any other dictionaries. All while 
intelligently sharing the decoded glyph data, of course.



git-svn-id: http://svn.ghostscript.com/jbig2dec/trunk@285 ded80894-8fb9-0310-811b-c03f3676ab4d