Replace the code that removes the additional CFLAGS added by the AC_PROG_CC autoconf macro. Closes bug 687150.
DETAILS:

On linux with gcc, AC_PROG_CC adds '-g -O2' to the general CFLAGS. This was overriding
the -O we add in Makefile.in, and the bug complains that this confuses gdb because of
code re-organization.

This change also means we no longer include full debug symbols in the normal build. I
don't consider this an improvement personally, but will bow to the traditionalists. Note
that 'strip' still reduces the size of the executable by 20%, so it's still worth
running if you're concerned about install-size.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@4369 a1074d23-0009-0410-80fe-cf8c14f379e6