Fix for 687020 and 687146.  The pdfwrite device is incorrectly converting CalRGB color spaces into ICC color spaces.  Fix #1.
Details:

As Raph pointed out in the Wednesday's support call, there are actually
two problems.

1.  We were not correctly recognizing that we can use a CalRGB color space
for the output.  This patch corrects this problem.  One of the tests in
that process was assuming an incorrect ordering to the transforms.  This
fix corrects this problem.

2.  The ICC color space which is being created is not equivalent to the
original CalRGB color space.  This results in differing colors between
the original file and the output PDF file.  This problem is still open.

Note:  I am leaving these bug reports still open since problem 2 is still
open.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@4382 a1074d23-0009-0410-80fe-cf8c14f379e6