Fix : an access to uninitialized data in zbfont.c .
DETAILS :

Type 11 fonts got an occasional indeterministic data to pdata->Encoding,
which were inaccurately checked for r_size() <= 256.

We're keeping the check "r_size(pfe) <= 256" for backward compatibility,
rather we're unclear about it's purpose. It shouldn't be harmful.

EXPECTED DIFFERENCES :

I did not detected any on Windows, but I believe that the old code was indeterministic.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@4403 a1074d23-0009-0410-80fe-cf8c14f379e6