Fix .gethalftone to prevent /undefinedresource or other error from being left in the $error dict. Seen with UnusualFontMatrix.ps when running under startjob encapsulation.
DETAILS:

Internal procedures like .gethalftone that want to use stopped but don't
propagate the error up should use .internalstopped since this doesn't
set /newerror in $error.

The file mentioned has a 'quit' following the showpage which when running
under job encapsulation (false 0 startjob) is redefined to do 'stop'.
This means that the /newerror flag in $error was being seen when the
'stop' after the showpage was executed.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@4449 a1074d23-0009-0410-80fe-cf8c14f379e6