Fix MSVC debug targets: add explicit makefile specification. Fix bug 687205.
DETAILS:
Debug targets have always been broken on Windows because nmake by
default looks for "MAKEFILE" and fails. Until recently, the default
target was built without the -O2 flag and could be used for debugging.
This patch adds the correct filename explicitly to the debug target
so it can be used instead.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@4460 a1074d23-0009-0410-80fe-cf8c14f379e6