Fix for fonts being displayed using .notdef character (empty boxes or other appearance). While the test file doesn't follow the PDF 1.3 or later manual recommendations for providing a [1,0] cmap for an embedded symbolic TT font, Adobe Acrobat Reader can display these characters. When the 'post' table doesn't provide the glyphs required by the /Differences list, try another cmap [3,0] to get the correct glyph (names and glyph #'s). Bug #686959 for customer #1060.
DETAILS:

The Adobe PDF 1.4 Reference Manual has some "hints" for producing portable
PDF's, but as far as reading PDF's that don't follow these hints, all that
is said (in section 5.5.5, p.333) is:

    Note: Some popular TrueType font programs contain incorrect encoding
    information.  Implementations of TrueType font interpreters have
    evolved heuristics for dealing with such problems; those heuristics
    are not described here.  For maximum portability, only well-formed
    TrueType font programs should be used in PDF files.

The approach of trusting the 'post' table unless the glyphs given do not
include glyphs that are mentioned in the 'prebuilt_encoding' (Differences),
is a heuristic that works for this file and doesn't interfere with any
other PDF's in the regression test. The thing about heuristics is that
the 'intelligence' they embody is no better than the understanding of the
best solution, so this may crop up again later (but we hope not).

Note that simply using the 00030000 cmap [3,0] format causes the regression
file 'test_multipage_prob.pdf' to fail with some .notdef characters.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@4529 a1074d23-0009-0410-80fe-cf8c14f379e6