Fix (pdfwrite) : A mangled CFF was written when an Encoding uses 256 glyphs.
DETAILS :

Bug 687217 "pdfwrite erases a certain type1 font when I used 'space' character"
Thanks to Inoue Koich and to Mr. Hirata for the useful information
in the Comment #4 of 687217.

The reason was an old "NYI" :
 	/*
 	 * The count of encoded characters is only a single byte, so we
 	 * have to use a supplement for the last character.
 	 */
 	/****** NYI ******/

This patch implements that.

EXPECTED DIFFERENCES :

None with old compartefiles.
mbb-with-space.ps to be added to comparefiles.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@4558 a1074d23-0009-0410-80fe-cf8c14f379e6