Fix for "687252 subfiledecode errors on includes greater than the size of a uint".  In spite of the bug name, the problem is not related to the actual size of the subfile or to the size of a uint. The actual problem is a missing required parameter:  EODCount. However Adobe once again ignores the missing parameter and what is specified in the PRLM-3rd.
DETAILS:

The given test file has a SubFileDecode filter without the required
EODCount parameter.

Ghostscript is rejecting the test file because the parameter is not
defined.  The RangeCheck error message is the result of the following
piece of code in zfilter.c:

  if ((code = dict_int_param(op, "EODCount", 0, max_int, -1, &count)) < 0)
	 return code;

This sets a default value of -1 for EODCount and then checks it
against the range 0 to max_int.  Thus when EODCount is not found
we get the rangecheck since -1 is outside the allowed range.

In spite of the fact that the EODCount is a required parameter,
Adobe Distiller 6.0, my CP1700PS printer, and CPSI 3010.105 print
this file without complaint.

This fix changes the default value to 0 to match Adobe and adds a
comment explaining why the code differs from the spec.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@4594 a1074d23-0009-0410-80fe-cf8c14f379e6