Fix (PDF interpreter) : CropBox computed wrongly.
DETAILS :

Bug 687244 "wrong interpretation of negative mediabox coordinates".
The bug title appears incorrect. Actually the problem is caused by
negative CropBox coordinates.

.pdfshowpage_Install translates the coordinate system
to provide positive coordinates, but pdfshowpage_finish did not
account the translation when writing the CropBox.
Therefore it the resulting CropBox was inconsistent with the painting
when the source CropBox uses negative coordinates.

EXPECTED DIFFERENCES :

A progression with vtm2k.pdf .

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@4597 a1074d23-0009-0410-80fe-cf8c14f379e6