Fix for Bug 687255  pcl output shifted to bottom of page.  Correct the order of the PCL commands at the start of page for the ljet devices.
DETAILS:

This problem surfaced when a previous fix 686815 was made.  That fix
corrected  a line of code which was setting the page seize in the output
file.

With the correct page size being specified in the output file, there were
some side effects.  Setting the page size resets several other items,
including the starting page X and Y offsets.

The fix for this problem consists of simply moving the setting of the page
size three lines earlier in the file.  Now the page size is set prior to
the page X and Y offsets, which are set as part of this line:

       fputs(page_init, prn_stream);

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@4627 a1074d23-0009-0410-80fe-cf8c14f379e6