Adapt makefile support for compiling in zlib to support v1.2.x rather than 1.1.x. Linking to a shared version is unaffected.
DETAILS:

The zlib source was re-arranged from the 1.1.4 to the 1.2.x releases.
While binary compatibility was maintained, and we link just fine to a
separately build dynamic zlib, our custom makefile support for compiling
in the library source no longer works.

Rather than add a 'ZLIB_VERSION' makefile variable and the attendent
configuration costs, we've simply migrated completely to the new (much
faster) 1.2.x series. 1.2.1 is now the minimum version to be used if you
wish to compile zlib as part of Ghostscript.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@4760 a1074d23-0009-0410-80fe-cf8c14f379e6