Initial implementation of the JPXDecode filter using the JasPer library for decoding the JPEG 2000 data. This version only implements a postscript-level filter; PDF support will follow in another commit.
DETAILS:

A new FEATURE_DEV, 'jpx.dev', is added to the makefile to enable
support. The jasper library must be available to link with in this case.
The library is autodetected in the autoconf build, but not yet enabled
on other platforms.

The implementation leaves a little to be desired. While JasPer is
robust, it provides no way to connect its internal stream architecture
to ours, so we end up buffering all the data. This seems to be the best
we can do without modifying the library.

We try to handle data from a number of colorspaces, but the spec is a
little vague on what is required. We convert YUV to RGB internally, but
do not handle sub-sampled planes.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@4796 a1074d23-0009-0410-80fe-cf8c14f379e6