Fix bug where changing the object tag was ignored unless the color was also changed.
Added gx_unset_dev_color to the setting of object tags.

This assumes the caller will be setting the device color.
It only unsets the color if the tag changes.

This will fix will be cosmetically changed later to move the
global BITAG and setting routines into a more appropriate place
like gxdcolor

git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@2086 06663e23-700e-0410-b217-a244a6096597