Add a simple C program that uses the jasper library to parse JPX (jpeg2k) image files and embed them in a PDF 1.5 document. Useful for testing as there are few producers available at the moment.
DETAILS:

We make no attempt to parse the required metadata (resolution and so on) ourselves. It would be
simple to add this so that the program can stand alone. On the other hand, jasper is capable of
opening a number of other image formats and encoding them to jpx streams. We don't do that either,
but it would also be easy to add support for so that JPXDecode versions of arbitrary images could
be embedded.

Because of this, it's possible to embed e.g. baseline jpeg data instead of a jpx stream if you
feed the program the wrong file. Caveat executor.

Based on the jpegtopdf program by Tor Andersson.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@4878 a1074d23-0009-0410-80fe-cf8c14f379e6