Fix (PDF interpreter) : CIDMap was wrongly created from CIDToGIDMap.
DETAILS :

CIDMap size was equal to CIDCount, which gs_ttf.ps sets to the number of glyphs.
With a non-Identity CIDToGIDMap it appears wrong,
causing some valid CIDs to be out of range and to print as a notdef character.

Bug 687351 "Embedded TTF font in PDF incorrectly displayed.".

EXPECTED DIFFERENCES :

Bug687351.pdf

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@4891 a1074d23-0009-0410-80fe-cf8c14f379e6