This change implements "687431 Add determining equivalent CMYK colors for spot colors".  This change currently only affects the psdcmyk device.
DETAILS:

For a description of the implementation see the comments at
the beginning of gsequivc.c.

In addition to the comments in gsequivc.c, the gs_separation
structure was changed to simplify it.  The gs_separation_info
structure was removed and the names filed moved into the
gs_separations structure.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@5056 a1074d23-0009-0410-80fe-cf8c14f379e6