Log Message: Fixes bug 686956 Display device does not support separations. Add support for separations to display device. This requires a change to the display callback structure. Support is maintained for clients using version 1 of the structure.
DETAILS:
Add support for a new color format, DISPLAY_COLORS_SEPARATION,
which supports CMYK and spot colors.  A new callback
display_separation() is used to tell the client about
the names and CMYK equivalents of each separation.
The callback structure version number is incremented to 2
to show that has changed.  The display devices checks
whether it is passed the older v1 structure or the newer
v2 structure, and only calls the new separation callback
if it was given a v2 structure.
Only 8-bit/pixel, up to 8 components and 64-bit depth are supported.
In the Windows and gtk+ clients, a subset of these 8 separations
can be selected for display.
The new code is based on the tiffsep device.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@5137 a1074d23-0009-0410-80fe-cf8c14f379e6