Fix for 687565 Incorrect equiv color for Orange in Altona test file. The altona test file Altona_Visual_bb_1v1_x3.pdf contains a spot color called Orange.  An incorrect equivalent CMYK color is being determined for this spot color.  The result is that the spot color is being displayed as black for the devices which handle spot colors and attempt to determine equivalent CMYK colors for the spot colors.  These are the psdcmyk, tiffsep, and display (with -dDisplayFormat=16#a0800) devices.
DETAILS:

The Orange spot color is used in two different DeviceN color spaces.  The
first color space has components named 'None'.  These components are used
as part of a special feature described in section 4.5 of the PDF 1.5 spec.
This feature uses the None components as a means for passing values to
the color space's alternate tint transform function.  In the example in
the spec., these values are an equivalent color in the alternate color
space.

Our logic for determining an equivalent CMYK color for a spot colorant
sets all color components except for the desired spot color to zero and
then uses the tint transform function to form an alternate color.  This
results in the equivalent CMYK color being black since the None values
contain the real information for the tint transform function.

This change consists of check for the presence of DeviceN color space
components named None.  If any are detected then the equivalent CMYK
color logic exits rather than determine an invalid color.  This is
sufficient for this case since the second color space which contains
the Orange spot color does not have the None components.  For the
general case, a solution needs to be found which samples colors given
actual values for the None components.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@5147 a1074d23-0009-0410-80fe-cf8c14f379e6