Fix (pdfwrite) : Discard the output file seekability requirement in the /PS pdfmark implementation.
DETAILS :

The ps2write requires length of streams to be written before the stream.
The implementation of /PS pdfmark wrote a length in a separate object after the stream.
Now /PS pdfmark is implemented with pdf_enter_substream,
which writes the length directly into the object dictionary.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@5191 a1074d23-0009-0410-80fe-cf8c14f379e6