fixed bug with iapi, removed gs_lib_finit()'s call of gs_malloc_release.
DETAILS:

The problem is that single instance and multi instance interfaces
both call gs_to_exit() with different meanings.

If gs_lib_init0 was called then it should malloc_release
single instance exit library means free all.

If the iapi interface was used then the memory
should not be freed since gs_lib_finit() and gs_to_exit()
are called when the interpreter is exited but there is a
seperate delete_instance call that should free the memory

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@5222 a1074d23-0009-0410-80fe-cf8c14f379e6