Fix bug 687038 that resulted in a crash if the initial device was not pdfwrite.
DETAILS:

This was almost fixed by recent changes to gs_pdfwr.ps and
gs_setpd.ps for bugs 687325, 687436 and 687458.

Unfourtunately, the initialization of parameters only occurred
if the initial device was pdfwrite. A second patch removing
the initial device constraint on the '1010' .schedule_init
process was committed and fixes the problem.

Note that there was a work-around that if -dPDFSETTINGS=... was
given on the command line, it also worked.

EXPECTED DIFFERENCES:

none, since we don't test for this usage.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@5224 a1074d23-0009-0410-80fe-cf8c14f379e6