Handle invalid PDF's that use 0.0 font scaling (Tf operator). Fixes bug 687606 for customer number 780.
DETAILS:

If the font scale is 0.0, then just use a very small value. This may result
in a small unwanted dot, but that is probably not an issue since it is
an invalid value. The better approach for 'invisible' text is to use
TextRenderingMode 3. The 'pdfformaterror' function is used to complain
about this so the problem might get fixed by the PDF Creator.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@5232 a1074d23-0009-0410-80fe-cf8c14f379e6