Implementation of 687496 transparency and CMYK support.
DETAILS:

This change extends our current implementation of the PDF 1.4
transparency to include blending in a DeviceCMYK color space.
There is also a partial implementation of a DeviceGray blending
space.  However I do not have a method of generating test data
for a DeviceGray blending space.  Currently the choice of a
blending color space is based upon the process color model of
the output device.

Note:  There two differences expected in the regression testing.
These are with ai2.pdf and dina3_watermark.pdf.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@5241 a1074d23-0009-0410-80fe-cf8c14f379e6