Close the font file after the font has been processed (loaded into VM). This prevents a problem with using up too many file handles detected on Solaris. Thanks to Alex Cherepanov for the patch. Bug #687066.
DETAILS:

The patch has no effect on the cache. There is no point of not closing
the file after the content has been copied to memory.

On GNU/Linux GS has 250K VMThreshold and 1024 handles. That's enough.
On Windows GS has 1M VNThreshold and 512 handles. This is still OK.
On Solaris GS has 1M VMThreshold and 256 handles. Handles can be used up
before garbage collection starts. The following program fails on Solaris.

(a)(w) file closefile
1000 { (a)(r) file pop } repeat

Dependence on VMThreshold indicates that GC is not activated by
open(2) failing with EMFILE . This is a real bug, the patch is
just a work around that we may need to fix later, but not likely to
be a problem.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@5261 a1074d23-0009-0410-80fe-cf8c14f379e6