Fix for 687607   Text missing with TextALphaBits=2 and tiffsep device.
DETAILS:

The TextAlphaBits and GraphicsAlphaBits device parameters were not working
for devices with a depth of more than 32 bits.  The fix involves replacing
the copy_alpha device procs in the 40, 48, 56, and 64 bit memory devices
with the gx_default_copy_alpha routine.  These devices previously had a
copy_alpha routine which did nothing.  This was a mistake from when these
devices were created as part of the DeviceN work.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@5276 a1074d23-0009-0410-80fe-cf8c14f379e6