Translate the name of the gs executable stored in the utility scripts at install-time on unix. Closes bug 686863.
DETAILS:

The autoconf build supports a --with-gs=<name> option for setting the
name of the Ghostscript executable, which is stored in the makefiles as
'$(GS)', but scripts such at ps2pdf that expect to make invoke
Ghostscript made no use of this variable, causing version skew at best,
complete failure of that popular utility at worst. We therefore now
rewrite the hard-coded name each time we install the utility scripts.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@5303 a1074d23-0009-0410-80fe-cf8c14f379e6