Add the LZWEncode filter that actually compresses using the LZW method, rather than the dummy encoder in slzwce.c that expanded the input, but that avoided the patent issues. Bug 687554.
Also fix a long standing issue in filter closing that failed to flush the
target stream after the final s_process_write_buf with 'last' == true.
Bug 224134.

DETAILS:

The LZWEncode filter was moved over from an older version of Ghostscript
that actually had a working LZWEncode.

The s_filter_close logic (discussed at length in bug 224134) was fixed
to be consistent with the specification and with Adobe. While the
target stream is not closed (per the specification unless CloseTarget is
true), the specification *DOES* call for an automatic 'flushfile' which
is supposed to be propagated down the filter chain.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@5379 a1074d23-0009-0410-80fe-cf8c14f379e6