Fix SEGV caused by premature free of the bbox device with high level devices that use the bbox_device of the vector high level device. Bug 687631.
DETAILS:

The vector device allocated a bbox device, but never incremented the
reference count past zero, so any use of the device that added a forwarding
device with that bbox_device as the target would free the bbox device
when that forwarding device closed. This happens with clipped images
or images that use ROP's since a clipping ROP forwarding could be used.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@5381 a1074d23-0009-0410-80fe-cf8c14f379e6