This is a partial fix for 687706 Display device 8-bit native incorrectly dithers grays.  Russell Lang is improving the 8-bit (256 color) display device to allow more shades of gray.  This change is required for that effort.
DETAILS:

The pre 8.00 release had more gray levels for the 8-bit display device
but the pre 8.00 release had other related problems in halftoning.  The
fix for those problems eliminated the extra gray levels.

Russell's effort dug up some problems with the handling of halftones
for devices in which the max_color and max_gray values are different
and gray_index is set to an actual component number.  This fix corrects
those problems.

Russell has a patch for the remainder of the effort.  His changes
involve changing the 256 color display device to use a DeviceRGBK process
color model:  red, green, blue and an additive black (gray).  New
device color mapping procs also have been created for this process color
model.  (This is the first example of this process color model in
Ghostscript.)

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@5396 a1074d23-0009-0410-80fe-cf8c14f379e6