Correct errors in .buildnativefontmap procedure.
DETAILS:
The enumeration of fonts returned by .getnativefonts
was being given "array true" which was causing a
typecheck in forall because of a missing exch.
It should have been "true array".

Also, .findfontname was returning a name rather than a
string, so cvn was failing because it was already a name.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@5535 a1074d23-0009-0410-80fe-cf8c14f379e6