Add a platform-specific JASPERCFLAGS makefile variable to assist with proper compilation of the jasper jpeg2k library. Set this variable correctly in the autoconf build.
DETAILS:

The jasper library natively supports building under both autoconf+cc and
MSVC, but expects a preprocessor define to be set to distinguish the
two. While we do carefully construct a special JAS_CC for compiling the
jasper source into Ghostscript, as with all the other 3rd party
libraries, there is only a facility to set special CFLAGS globally.

This change adds a new top-level makefile variable for passing
platform-specific CFLAGS into gs.mak for addition to that global
set.

Also, set JASPERCFLAGS=-DJAS_CONFIGURE in Makefile.in so that the jasper
sourcr expects the autoconf generated configuration header when
compiling as part of Ghostscript.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@5545 a1074d23-0009-0410-80fe-cf8c14f379e6