Add JPXDecode filter to the default Windows MSVC build. This is based on the "jasper" source and must be at least version 26 from the svn repository using svn co http://svn.ghostscript.com:8080/jasper/trunk jasper
DETAILS:

This is a fairly minor modification to the gs MSVC makefiles, but it did
require changing our jasper version (that had accumulated many changes)
to hook in the correct jas_config***.h file and to avoid using windows.h
(so we can build with /Za). See the jasper svn for details of the changes.

EXPECTED DIFFERENCES:

none. (we didn't have any PDF 1.5 files with JPEG2000 images in the
comparefiles).

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@5547 a1074d23-0009-0410-80fe-cf8c14f379e6