Convert gs_malloc() calls in the MacOS gp_* code to unwrapped libc malloc() now that gs_malloc() requires a memory pointer, which is not available.
Partial fix for bug 687830.

DETAILS:

Only the call in getenv() is active; the one in gp_init() is not used.

This may have been what the earlier libctx patch was trying to do when
it touched gp_macio.c, changing the function call, but not the arguments,
and breaking the

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@5562 a1074d23-0009-0410-80fe-cf8c14f379e6