Provide more robust handling of errors in an attempt to prevent badly formed error handlers defined in the PS file from causing nested errors (and the infamous dump-like "Unexpected interpreter error"). Refer to bug 687925.
DETAILS:

Note that broken error handlers are fairly common, so running them in a
.internalstopped context makes sense. The .internalstopped is used to
avoid side effects such as modifying the $error dict contents.

Also note that the -dJOBSERVER may often be needed to actually see the
error message since this forces the use of the "standard" error reporting
to stdout.

Expected Differences:

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@5679 a1074d23-0009-0410-80fe-cf8c14f379e6