Eliminate a performance problem when a TT font has a large 'loca' table. A fix for 687889 used .sort on the 'loca' table but this takes a long time if the table is large. Only run the .sort if the table is out of order which is rare since the TT font spec requires an in-order table. Bug 687968.
DETAILS:

Apparently our conversion of the 'japan.ps' file creates a TT font with
a loca table that has 20458 entries (even though the original font
in japan.ps didn't have that many). This is needed because there are
many glyphs that are not present, but we need loca table slots that
represent the missing glyphs (same starting offset as next glyph).

The time was being spent doing the '{ lt } .sort' of the array  :-(

I added logic that checked the order as the loca table was converted
to an array so that the .sort could be skipped as long as the table is
in order (which it *should* be for a valid TT font). After this change,
the HEAD time is 1.16 CPU seconds versus 1.08 seconds for gs8.50.
(with the .sort, the time was 257 seconds).

I suppose that we could optimize the .sort (put it into 'C') but
for now I'll just avoid it most of the time.

Note that since the .sort has a procedure callout for the ordering test,
it might not speed up that much in 'C'. That's the other reason that it
is worthwhile to avoid the .sort when possible.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@5723 a1074d23-0009-0410-80fe-cf8c14f379e6