Set the default paper size with DEFAULTPAPERSIZE in gs_init.ps, which only affects devices that default to A4 or letter paper. This avoids inappropriately setting the bbox device to A4 or letter size.
DETAILS:
Some linux distributions change the default paper size
to A4 by uncommenting the following line in gs_init.ps.
  % Optionally choose a default paper size other than U.S. letter.
  % (a4) /PAPERSIZE where { pop pop } { /PAPERSIZE exch def } ifelse

Unfortunately this sets the default paper size for all devices,
including the bbox device for which this is not appropriate.
If the bbox device is set to A4 paper and drawing occurs
outside this area, the bbox device will report an incorrect
bounding box.  There may be custom printer devices which don't
support A4 or letter paper.

The default paper size is now set with DEFAULTPAPERSIZE,
which sets PAPERSIZE only if the paper size is not otherwise
set, the device defaulted to A4 or letter.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@5733 a1074d23-0009-0410-80fe-cf8c14f379e6