Add FillOrder device parameter to TIFF fax devices and document the changes in Devices.htm . Fix bug 687324
DETAILS :
This fix changes the default from FillOrder = 2 to FillOrder = 1.
According to the TIFF 6.0 specification, Section 8, page 32, support of
FillOrder = 2 is not required in a Baseline TIFF compliant reader.

Support of FillOrder will be addressed to tiffpack and tifflzw devices as
a part of bug 614298 resolution.

EXPECTED DIFFERENCES :
None. TIFF devices are not tested by the regression test.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@5744 a1074d23-0009-0410-80fe-cf8c14f379e6