Fix for 688011 Segment fault with -r150 and the mswindll device.
DETAILS:

The routine gx_image_cached_char creates a mono bit memory device with
the memory structure pointer set to 0.  It then passes this device to
win_render_char which seg faults when it tries to allocate memory for
a bit map.

The fix consists of using the parent device's memory structure for the
mono bit memory device.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@5811 a1074d23-0009-0410-80fe-cf8c14f379e6