Fixed bug in pcl top margin.
Setting a fixed CAP to top of form was occuring on an unmarked page.

Moving CAP should make the cap fixed otherwise it is floating.
Rather than mark cap as fixed/floating I'm cheating and looking
at its position relative to margins to denote fixed.  This in
combination with is marked is used to control
moving the cap to top of form on this page or the next.

It is possible fool this code with the right combination of multiple
cap moves and top margin calls before marking the page.

Tradeoff between fool proof and performance code turmoil.

git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@2231 06663e23-700e-0410-b217-a244a6096597