Fix: use identity matrix (rather than scaling by font matrix) to prepare parameters for CDevProc procedure in pdfwrite code path. Fixes bug #688033.
DETAILS

The problem in the 688033 test file was that the CDevProc procedure
for the font was getting called with values that were already scaled
down by .001, even though the PLRM3 says "all of the numeric values
are expressed in the glyph coordinate system of the font." (5.9.2,
p. 350)

My proposed fix is to pass in the identity matrix, instead of the font
matrix, to the gs_default_glyph_info procedure. This patch was tested
with both the 688033 test file and the test file from 687681, which was
the bug that motivated the changes to pdfwrite CDevProc handling. The
patched code also passed a full local regression.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@5844 a1074d23-0009-0410-80fe-cf8c14f379e6