Prevent out of range ImageType 4 mask array entries from causing an error as Adobe Acrobat Reader does. Bug 687611 for customer 870.
DETAILS:

When the Mask array values are beyond the max allowed by the BitsPerComponent,
a rangecheck error must be generated by PostScript, but Adobe Acrobat Reader
ignores the Mask in this case and effectively treats the image as ImageType 1.
This patch generates a pdfformaterror Warning when the problem is seen.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@5860 a1074d23-0009-0410-80fe-cf8c14f379e6