Fix for 688010 Errors in overprint with halftoned output.
DETAILS:

The PDF overprintmode can be used to specify overprinting when the  value
of an input color is zero.  The problem was caused by an error in
determining when an input color was zero.  The problem was only apparent
with a 'binary' halftone.  Ie.e one in which on a single component is
being halftoned.  The earlier logic would overprint all components which
were not being halftoned.  The example case involved 100% black and 50%
cyan.  The black component was treated like it was 0%.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@5922 a1074d23-0009-0410-80fe-cf8c14f379e6