Fix VM space for userparams. Problem and patch courtesy of Paul Vojta. Bug 688107.
DETAILS:

Also thanks to Alex Cherepanov for a reporducible test case. The original
problem file 'epstest.eps' did not actually show a problem. I tested this
against the comparefiles to validate that there were no unexpected changes,
but generally it was an obvious problem that an array might be created
in a VM space that was not the same local/gloabl as the strings it contained.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@5924 a1074d23-0009-0410-80fe-cf8c14f379e6