Fix for 688113 null ptr crash in gdevp14.c.  With the -rXXXxYYY switch specified and high resolution PDF 1.4 transparency files, GS would crash.
DETAILS:

The logic for deteriming band sizes was being bypassed if the -rXXXxYYY
switch was specified on the command line.

Specifying resolution on the command line was forcing the calculation of
band size.  This occured before the determination that PDF 1.4 transparency
was being made.  The band size was not being recalculated when the
determination was being done.  The fix consists of including the
'page_uses_transparency' as part of the logic for determining when the
band size is recalculated.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@5927 a1074d23-0009-0410-80fe-cf8c14f379e6