Fix to allow CCITT data without EOFB when EndOfBlock is true (the default) to return normal end-of-file (EOFC) status instead of ioerror (ERRC) as long as the row count given by Rows is satisfied. Bug 687973.
DETAILS:

This bug was reported by a potential customer. The fix is minimal and
returns EOFC instead of ERRC if the count given by Rows has been read.
This seems to be the closest to what is implied by the Adobe description
about EndOfBlock in that the filter WILL continue to process data past
the Rows value if EndOfBlock is true (the default) as long as no error
is seen. This means that PostScript programs that depend on this action to
consume extra data will still operate correctly.

Expected Differences:

None (I ran all of the comparefiles).

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@5937 a1074d23-0009-0410-80fe-cf8c14f379e6