Fix for the BeginPage and EndPage procedures used for N-up processing.
DETAILS:

The BeginPage called the previous BeginPage (stored in .BP) at the wrong
time, so that rotation/translation of individual pages would not work
correctly.

The EndPage incremented the .Pn value even on 'device deactivation' calls
(mode == 2) so the page count was off if the input PS did 'setpagedevice'.

This probably is not widely used since it was broken. It's a little better
now, but still needs improvement (IMHO) such as setting a 'clip' before
the indivual pages. Try running two examples/tiger.eps to see why.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@5943 a1074d23-0009-0410-80fe-cf8c14f379e6