Add support for manual feed and media source or tray selection to pxl output devices.  This allows postscript input to use InputAttributes dictionary selection mechanisms to choose a tray.
The pxl tray numbers in the PXL file are the PS MediaSource numbers
to "map" correctly an input attributes dictionary needs to be present
and needs to match the tray contents/numbering of the pxl printer.

postscript manual feed /ManualFeed is directly mapped to PXL this
will override any tray selection.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@5995 a1074d23-0009-0410-80fe-cf8c14f379e6