Remove the arbitrary limit of 256 max_colors from the dci_ macros. This is related to bug 688204 and allows shadings (gradients) to have more than 256 steps for high precision color devices.
DETAILS:

Note that the default smoothness is set to 0.02 in lib/gs_ll3.ps and unless
this is set smaller with the setsmoothness operator (or the PDF 'SM' op),
the max_error used in src/gxshade.c may still be too large to see REALLY
smooth gradients. The next part of the changes is to modify the setting
of max_error when the device has > 256 levels.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6012 a1074d23-0009-0410-80fe-cf8c14f379e6