Fix handling of Some Type1C (CFF) font subsets where the tables may not be completely contiguous. Seen with PDFs created by pdflib. Bug 688216 and 688040 for customer 562.
DETAILS:

The example subset CFF had the "Private" data about 30,000 bytes after
the end of the previous (CharStrings) data. While we try to process the
CFF file sequentially, there is no reason we cannot skip bytes. If we
encounter a CFF where positioning (backwards) is required, we will have
to further change the code to set us a "ResuableStreamDecode" file so
we can replace the "pos sub nextstring pop" with "f fileposition" which
will work forwards or backwards.

A CFF that requires this will get an error when loading the font from the
-string- operator (rangecheck). Since we have not seen one yet, we will
wait until this surfaces.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6013 a1074d23-0009-0410-80fe-cf8c14f379e6