Fix handling of paths so that 'cm' will affect pending path operations (fill stroke or clip). Verified that with this patch we do as Acrobat Reader from 4 through 7 does. Bug 688299 for customer 870.
DETAILS:

The basics of using 'upath', then 'uappend' to capture, then replay the
path after the 'concat' done by 'cm' is straightforward, but since the path
may not be used, there may be coordinates that are WAY out-of-range after
an arbitrary 'cm' so we execute the 'uappend' inside a 'stopped' context.

EXPECTED DIFFERENCES:

None. I did a check of all PDF's in the regression suite -- none had
any differences.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6119 a1074d23-0009-0410-80fe-cf8c14f379e6