Fix for 688344 Error: /invalidaccess in --setfileposition-- after finding bad XREF.
DETAILS:

The test file contains xref table entries that have locations beyond the
end of the file.  When the xref table entries were being verified, this
would cause an attempt to seek beyond the end of the file.  This closed
the file.  The attempt to rebuild the xref table then crashed.

The fix consists of adding a check for xref entry locations being within
the file.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6164 a1074d23-0009-0410-80fe-cf8c14f379e6