Make sure procedure target of an Encode filter gets called with the end of data condition, i.e., 'false' condition. Bug 688326.
DETAILS:

When a procedure is a data target, it is signalled that it has been closed by
one final callback with a possibly empty string containing any final buffered
data, and the boolean false.

When 'last' was true if the count was 0, the previous code would not return
CALLC status which is needed to call the procedure with the 'false' status.
If ss->eof is already true, then the call has already been done.

This condition is not checked in any of the Genoa test suites.

EXPECTED DIFFERENCES:

none.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6168 a1074d23-0009-0410-80fe-cf8c14f379e6