pdfwrite : Generate document Metadata.
THIS IS INCOMPATIBLE CHANGE.

Adobe Distiller 5 generates metadata according to Adobe XMP speification
iff the distiller parameter ParseDSCCommentsForDocInfo or PreserveEPSInfo is true.
Old versions of Ghostscript does not do that.
New versions of Ghostscript do when CompatibilityLevel >= 1.4 (default).
Generated PDF documents become longer in about 1.5 kilobytes.

See the documentation change about new command line options and distiller parameters.

DETAILS :

It's a part of the PDF/A project.

This is an initial commit for the subject.
It generates Document metadata as Adobe Distiller 6 does
for a simple file (Genoa test 001-01.ps).

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6172 a1074d23-0009-0410-80fe-cf8c14f379e6