Insert a space character before "%%" or "%!" in ASCII85Encode filter to avoid confusion with DSC comments. Fix bug 688079
DETAILS :
ASCII85Encode filter has a special treatment of some output lines that looks
like DSC line, (ie. lines that begins with %% or %! are broken after first
percent). This didn't work when the last line of encoded stream begins with %%
and input stream was ended not on full 4-tuple.

EXPECTED DIFFERENCES :
None

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6185 a1074d23-0009-0410-80fe-cf8c14f379e6