Add capability to display text Value strings in Widget Annotations. Multi line support is not yet implemented, but basic Tx types are supported with /DA appearance string (used to select font and color, etc.) as well as /Q (quadding == justinfication) properly supported. Bug 688368 for customer #670.
DETAILS:

This is an initial implementation sufficient for the customer's immediate
requirement. NYI includes handling Flag (/F and /Ff) bits, one of which
is supposed to make the rendering "invisible".

Note that if the /AP keyword is present, that appearance will be used to
render the Annotation rather than the /V value based rendering. This was
needed to properly render the 01_001.pdf where both keys are present.
Presumably, the /AP appearance was 'baked in' by the form filling application.

Also note that the /DA string in the sample file used "/FontName scale Tf"
which required making the /Tf implementation handle both a font dictionary
(usual case) as well as a font name (pdf_ops.ps change).

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6188 a1074d23-0009-0410-80fe-cf8c14f379e6