Fix (font machinery) : Provide an additional debug data about text enumeration.
DETAILS :

This relates to Bug 688422 "Assertion failed! in gxccman.c".
It doesn't fix the bug, simplifies its analyzis.
We decided to make this change permanent because
in the past we spent many efforts for anayzis of such kind.

Rather text_enum_id is not used with release build,
it is still defined to keep same structure size.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6211 a1074d23-0009-0410-80fe-cf8c14f379e6