Correct a few typos in 16-bit pixel differencing filter. Bug 687986
DETAILS :
This is a partial fix for a dead code. The patch fixes simple typos in 16-bit
code, which are difficult to spot because of the nested macroses.  Assignment
ss->prev[0] = s0 is needed because the code for handling leftover bytes uses
ss->prev[0]. The filter still falls into an infinite loop sometimes, but this
fix doesn't address this problem.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6238 a1074d23-0009-0410-80fe-cf8c14f379e6