Fix handling of broken TrueType fonts that have a loca table that is not in order (continued).
DETAILS :

Bug 688461 "TrueType-Problem with Core-Dump (Solaris)".

This is a further improvement of the patch for
Bug 687889  "/rangecheck in --string--".

1. Inserted a code for ignoring excessive data at end of 'loca' - see comment in code.
2. When (1) works, a warning is printed to stderr (in debug build only).
3. Fixed C coding style in gstype42.c revision 1.49.

We're not sure that this patch is good for any incorrect TrueType font,
because the old code keeps some artifacts without enough explanation.
We inserted a debug printing to study related cases better.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6241 a1074d23-0009-0410-80fe-cf8c14f379e6