Enable use of "Bold" in BaseFont name when doing font substitution from PDF files that don't embed all fonts. Bug 688432 for customer 670.
DETAILS:

There is not really a 'Bold' flag in the FontDescriptor Flags (the ForceBold
infers a Bold font, but is not present for all Bold fonts since it has a
different meaning).

The pdf_font.ps logic had logic for "Narrow" in the font name, but not "Bold".
Enabling recognition of "Bold" works for this file and probably most others.

EXPECTED DIFFERENCES.

unknown, but if so progressions will be updated.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6244 a1074d23-0009-0410-80fe-cf8c14f379e6