Remove the check for files not listed in CVS.
DETAILS:

This no longer works since we've switched to subversion for version control. The
check could of course be ported, but the object was the detect files that had
either accidentally been left out of a commit, or files that might slip into a
release unintentionally. The 'svn status' command is an effective and more
commonly used check for the former, and our policy of building releases from
a clean export of the tag eliminates the other.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6441 a1074d23-0009-0410-80fe-cf8c14f379e6