Fix for 688488 Since Jan 7 2006, many regression files get 'typecheck in --execfunction--'
DETAILS:

The commit for 688413 introduced a new problem.  If the attributes dictionary contains a reference to a separation color space which had already been resolved, then a conflict occured with the processing of the alternate tint transform function.  This patch eliminates the conflict by creating a new copy of the DeviceN color space array and the attributes array when the color space is resolved.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6446 a1074d23-0009-0410-80fe-cf8c14f379e6