Update the PDF text matrix basis (TextSaveMatrix) when a 'cm' occurs within a BT/ET section. Bug 688485 for customer #850.
DETAILS:

Apparently this is rare -- none of the previous regression tests had this case
(a 'cm' inside a BT/ET section). The 'missing' Arabic text was actually outside
the clipping path, so nothing was painted.

EXPECTED DIFFERENCES:

None. I ran the PDF subset of the regression suite and none of the files had
this.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6458 a1074d23-0009-0410-80fe-cf8c14f379e6