Make sure that the final gs_closedevice isn't called on a device that has been freed during the final .uninstallpagedevice. Bug 688494.
DETAILS:

The overprint device or other compositors can (under some circumstances) have
a reference count of 1 when the gs_lib_finit is called, which means that the
'nulldevice' operator executed by the .uninstallpagedevice will free the
compositor when the rc_assign to the graphics state is performed.

To avoid an operation on a pointer to freed memory, don't attempt to close
the device if the ref_count prior to the unistallpagedevice was 1.

EXPECTED DIFFERENCES:

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6461 a1074d23-0009-0410-80fe-cf8c14f379e6