Fix : A wrong CTM in clist after a create compositor command.
DETAILS :

When a create compositor command is written to clist,
the imager state copy was not updated in the clist writer device,
causing a further CTM desyncronization.
It caused a wrong clist playback for the bands affected by the compositor.

Debugged with GS_Stroke_error.pdf, bugsample.pdf, Bug687724.pdf
with the patch for 688518 applied.
Have no tests for observibg this bug without the patch for 688518.

EXPECTED DIFFERENCES :

None,

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6536 a1074d23-0009-0410-80fe-cf8c14f379e6