Removes hexadecimal conversion from the code and pass the encrypted part of the type 1 font to eexec operator intact in all cases. Bug 688385 from customer 1110.
DETAILS :
The PDF specification states that Type 1 fonts always use binary form of
the encrypted part, but some PDF producers embed fonts with hexadecimal
encoding unchanged. The standard eexec operator recognizes and decodes
them transparently.

The only case that is not handled is a binary-encoded file that looks like a
hexadecimal encoding to eexec. This case has not been observed in the wild.

EXPECTED DIFFERENCES :
None

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6568 a1074d23-0009-0410-80fe-cf8c14f379e6