Calculate the last byte of the character code modulo 256 to prevent the final value exceed 256 and cause an rangecheck error. Bug 688535.
DETAILS:
The overflow and the proposed adjustment can happen only when the
last element is 0xFF.

The last value of the string is not used but discarded at the end of the loop.
So the adjustment has no effect other than suppressing the error with
the minimal changes in the existing code.

EXPECTED DIFFERENCES:
None

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6570 a1074d23-0009-0410-80fe-cf8c14f379e6