Clamp bbox rectangle to avoid q<p case that the rect_intersect macro allows to avoid allocating transparency buffer that has an incorrect (way too large) size. Bug 688546.
DETAILS:

WHile it might be more elegant to skip a lot of the transparency logic, it
works to simply end up with a zero width or height of the rectangle. Since
this is an anomalous case, optimization is not that important.

EXPECTED DIFFERENCES:

None. This case did not exist previously in the regression suite.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6573 a1074d23-0009-0410-80fe-cf8c14f379e6