Implement yet another hack for an invalid PDF file that Adobe handles. If a parameter is missing from the Page tree, check the Root Catalog. This case had MediaBox in the Root dict, but not anywhere in the Pages tree. Bug 688562 for customer 670.
DETAILS:

MicroPatent probably misinterprets the spec where it mentions "For example,
a document might specify the same media box for all of its pages by including
a MediaBox entry in the root node of the page tree." thinking that the Root
Catalog is the Root of the Page tree, rather than the /Pages object from the
Root.

As is usual, I issue a pdfformaterror Warning when this is detected.

EXPECTED DIFFERENCES:

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6594 a1074d23-0009-0410-80fe-cf8c14f379e6