draw gradients inside a tiling pattern (that should not tile) so that it will work with henrys path code. this steps on met_pattern_t which was only used for images, so there are some changes in metimage.c
leaks like a sieve. the CT_*GradientBrush structs are kept around in
the met_pattern_t which is kept in the gs_client_pattern. i also strdup
the strings in the CT_structs so that expat won't step on them, since they
will hang around after the _done is called.

git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@2328 06663e23-700e-0410-b217-a244a6096597