Fix a problem with CMYK JPXDecode filter image streams.
DETAILS:

The Luratech jp2 decoder does not return a general image size for CMYK
images (this seems to be a bug) or for images with subsampled
components. Therefore we use the dimensions of the zeroth component
rather than querying the overall image (component -1).

When the JPXDecode filter implementation is extended to pass the
required Colorspace information from the image dictionary, it should
also pass the specified width and height of the image, since this
is the actual specification the filter must produce data for.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6633 a1074d23-0009-0410-80fe-cf8c14f379e6