Fix the unix install to include Fontmap.GS. This was the cause of 688567.
DETAILS:

The mkromfs utility uses the passed paths as templates for file
enumeration, so EXTRA_INIT_FILES was added containing only
'Fontmap' and 'FAPI' as a shorthand for the collections "Fontmap*"
and "FAPI*". However, this variable is also used in unixinst.mak
as a literal list of files to install, and so only 'Fontmap' was
actually being installed.

This fix remains confusing, but in practice handles the install.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6636 a1074d23-0009-0410-80fe-cf8c14f379e6