Fix luratech-based encode filters. We now appear to work correctly.
DETAILS:

Invert the JBIG2 image data before passing it to the encoder library. We
allocate a scratch buffer for this because the buffer we are passed in
the read cursor struct is marked const. We now produce output with the
correct meaning for 0 and 1 bits, but the output of the jbig2 output
device is now inverted.

In both the jbig2 and jpx encode filters, repeatedly grow dynamic
buffers until they reach the required size instead of assuming a
single doubling will be sufficient. This fixes a segfault inside
the library compressing large JPX images.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6684 a1074d23-0009-0410-80fe-cf8c14f379e6