Work around a bug in GCC 2.96 for x86. Pre-load stack pages to avoid a SEGV. Bug 688574.
DETAILS:
It is well known that GCC 2.96 for x86 sometimes forgets to adjust $esp
and leaves automatic variables at small distance below the stack pointer.
Apparently, when the access to the automatic variable causes a page fault
Linux sends a SEGV signal instead of committing the page if the access
happens below the stack pointer. Pre-loading the stack pages from a known
good context resolves the problem.

DIFFERENCES:
None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6697 a1074d23-0009-0410-80fe-cf8c14f379e6