Fix so that dictionaries that grow preserve attributes. Several problems reported including a potential security hole when in SAFER mode. Thanks to mat Gh Savulescu (SaGS) for the bug report, the patch and excellent analysis of the problem.
DETAILS:

Ordinarily dictionaries have unprotected, non-exectue
status, but it is possible to "grow" executable or even
"read-only" dictionaries (contact Ray Johnston for an
explanation as provided by mat Gh Savulescu). The bug
in idict.c would not preserve the attributes, making
the dictionary not executable, and not readonly even
if those were the original attributes.

EXPECTED DIFFERENCES:

None. (we don't run regression in SAFER mode and don't
rely on 'special' dict attributes).

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6765 a1074d23-0009-0410-80fe-cf8c14f379e6