Fix (CIDFontType 1) : Check for GDBytes elements to point past the 'loca' table end.
DETAILS :

Bug 688669 "/invalidfont in --@0x100194e0--".

The old code didn't check GDBytes consistency, taking occasional data
past the 'loca' table with an occasional result.

Note that the change to z11_CIDMap_proc in zfcid1.c trows exception,
which is handled by the procedure %Type11BuildGlyph
in gs_cidfn.ps with substituting .notdef .

The test document embeds an incorrect True Type data for the font OBMKGG+H2gtrM.
The True Type header specifies the size of 'loca' table equals to 12 bytes.
In the sfnts array it is a separate string with the array index 2.
Then it renders the glyph index 3 accessing outside the loca table.
It hapens while rendering the text :

/TT5 1 Tf
0.365 0 TD
<0001>Tj

Probably the problem happens because 2 font descriptors (43 0 obj and 38 0 obj,
which correspond to local font names /TT4 and /TT5) refer to same single font data 39 0 obj.
This data is an embedded font subset, which is enough for /TT4 and has no
glyphs for /TT5. This problem to be reported to Adobe Distiler 6,
which created the test document. Is the version 7 better ?
(can't check due to no Postscript source).

In the test case .notdef is the CID 0 and defined in the font as a hollow box.
Don't know why Adobe doesn't print it in the test document.
That's another story. The font doesn't define a space character.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6874 a1074d23-0009-0410-80fe-cf8c14f379e6