Increase the operand stack size limit to the current working maximum = 65414. Bug 688764.
DETAILS:
PostScript stack cannot exceed the size of a PostScript array (65535) because
the stack is copied to an array when the stack overflows.

The stack space is allocated in blocks. The stack overflow is checked when the
new block is allocated. If the MaxOpStack is set close enough to 65535 the
actual stack size exceeds 65535 at the moment of stack overflow.

This condition is detected when the stack is copied to an array during
stackoverflow processing, causing a low level stack dump.

DIFFERENCES:
None

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6889 a1074d23-0009-0410-80fe-cf8c14f379e6