Fix (pdfwrite) : More PDF/A-1b complience, continued 2.
DETAILS :

Bug 688742 "PDF/A label missing error in Acrobat preflight", part 4.

This is an experimental patch, which drops the setting of
the symbolic flag for fonts in PDF/A documents.
Doing so because the Acrobat preflight tool
doesn't allow Encoding for symbolic True Type fonts.

The decision for setting the symbolic flag for most fonts has been
done a long ago, and currently we cannot reconstruct all the related history.
We suspect that in novadays that decision appears obsolete.
We commit this patch to allow an accumulation of information
about how this flag affects 3d party software.

DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6895 a1074d23-0009-0410-80fe-cf8c14f379e6