Fix (pdfwrite) : Support OS/2 table version 2 in True Type fonts.
DETASILS :

We don't use the new fields in the 96-byte structure and continue to
create 86-byre structures. The extra space is used as a buffer when
we copy the structure intact.

Bug 688536 "/invalidfont in --show-- when trying to convert the PDF file".
Patch from Alex with an improvement against regressions.

DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6919 a1074d23-0009-0410-80fe-cf8c14f379e6