Many pcl commands were not properly "locked out" when the pcl interpreter was in graphics mode.  Progressions noted in the following CET files:
16-08.BIN
21-03.BIN
21-09.BIN
26-01.BIN
35-05.BIN
35-08.BIN

5c fts 0880 panel 882, a difficult one to spot, the raster was not
positioned correctly in the black box.

Also fixes missing wrapped text in CET File:

16_02.BIN

pcpage.c: PCL was setting the color to solid white upon erasing the
page which is usually harmless because the color is reset before the
next graphic.  However, in text wrapped from one page to the next the
color reset did not take place resulting in white text.

git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@2482 06663e23-700e-0410-b217-a244a6096597