Set the correct error object when special operators (x_add, etc.) fail.
DETAILS:
If control reaches the special operators (x_add, etc.) as a result of
interpreting an executable name, iref points to the name, not the
operator, so the name rather than the operator becomes the error object,
which is wrong.  We detect and handle this case explicitly when an error
occurs, so as not to slow down the non-error case.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6958 a1074d23-0009-0410-80fe-cf8c14f379e6