Fix (pdfwrite) : Wrong xref : Unuseful object ID when a font is not really used in the document.
DETAILS :

Bug 688824 "invalid xref table from pdfwrite",

The old code reserved an object ID for the FontFile object
when the font is set but no text rendered with the font in the document.
The new code skips creating the FontFile object if
no object ID assigned to the font descriptor.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@6983 a1074d23-0009-0410-80fe-cf8c14f379e6