WARNING 1: This patch modifies the urw font set so you must recopy/rewrite the fonts in your font directory.
WARNING 2: This change requires an associated change by Monotype if
building with the UFST.  We are waiting for that change.

plchar.c: removes unnecessary debug warning.

pcparse.c: Parser was throwing away ESC and the following character, if
the second character did not form a known command, only the ESC should
be discarded.  Fixes missing data in CET 18-14.

pctext.c, pglabel.c and urwfonts: Removes the space character from the
proportional fonts to more closely match the hp resident set.  The
"magic" in pcl to determine space character vs control code did not
always work properly.  It is much simpler to get rid of the space in
the font and this is consistent with the fonts in the UFST.  Improves
cet 20-01, 20-02, 20-04.  In all of these cases a downloaded symbol
set had redirected the space slot to another character and we
interpreted this as a space.

REGRESSIONS: There are minor regressions in the Extra Space test in
the CET but this code is being worked on now anyway.

git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@2534 06663e23-700e-0410-b217-a244a6096597