Add more pcl command lockouts in raster mode. No regression test files exercise this yet.
Change the default symbol set logic.  The default symbol set is changable
via PJL and the printer has a default but this is just a suggestion
if a matching symbol set is not found the actual default
is PCL going back to its roots Roman 8.  When no symbol set is specified
the PJL default is used, if a requested symbol set is not found, Roman8.

Progressions for:

cet 30-03 pg 6,7, 8, 9, 12

Fixes:

cfts.0453
cfts.1670
efts.1340
efts.2063
efts.2330

git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@2535 06663e23-700e-0410-b217-a244a6096597