Port jpeg device scale and translate functions from GhostPCL.
DETAILS:

This patch adds four new parameters to the jpeg devices. In addition to
Quality, one can now set float parameters ViewScaleX, ViewScaleY,
ViewTransX, and ViewTransY so set a scale and offset to use when
rasterizing the page.

This patch also expands some of the device parameter macros and cleans
up the comments.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@7006 a1074d23-0009-0410-80fe-cf8c14f379e6