Fix for 'invalid restore' in CET test file 16-04.ps.
DETAILS:

There are were complications.  1)  The procedure .completfont was not properly
signalling an error when it occured.  2)  This procedure is used from both
definefont and findfont.  We need the error to be indicated as coming from
the appropriate procedure.

The fix for this problem is to fix .completefont to properly report errors
and to add stopped contexts around each of the invocations of .completefont,
loading the errorname from .completefont and calling signalerror again with
the appropriate procedure name.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@7017 a1074d23-0009-0410-80fe-cf8c14f379e6