Fix (ps2write) : Didn't write the OPDFRead procset.
DETAILS :

This patch includes 3 things :

1. In gdevpdfd.c : A temporary fix for Bug 688856
"Problems with ps2write (see comment #5)".
We keep the bug open because it still requires our
attention about the order of initialization events.

2. In gdevpdtt.c : Allow writing True Types with glyph indices over 255.
The old code is compatible to a bug in Adobe reader 3,
which is irrelevant to ps2write. The new code
keeps the old logic with pdfwrite and skips the check
with ps2write.

3. In opdfread.ps : .max was undefined.

This patch helps to work on the customer bug 688790
"pdfwrite : The PDF/A conversion needs to to re-encode texts"

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@7020 a1074d23-0009-0410-80fe-cf8c14f379e6