Merge the gs853merge branch down into the trunk. This ports ghostpcl and ghostxps to build against a branch of ghostscript close it its current mainline (around r7000) between Ghostscript 8.54 and 8.55.
This mostly consists of removing the memory pointers from many calls as
well as a few other api changes, as well as some makefile updates, which
were skewed between the branches.

Nevertheless this is a big merge and there are likely to be a few
mistakes. The previous trunk is tagged as /tags/ghostpcl-pre855merge
for comparision.

Both pcl6 and pspcl6 build and run with the Artifex font scaler (the
native gs one). pcl6 compiles but does not link with ufst. pspcl6 does
not compile with ufst, but that's not too different from what was
happening in the branch. Fixing this will of course come next.

git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@2566 06663e23-700e-0410-b217-a244a6096597