Fix : pfont->uid.xvalues uses a wrong allocator
DETAILS :

Bug 688883 "pfont->uid.xvalues uses a wrong allocator".

In the old code build_gs_simple_font allocates uid in the current memory,
but few moments later build_gs_sub_font changes the current space to the space
of the font dictionary, and allocates the font structure in it.

The new code allocates the font's uid in same space as the font dictionary.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@7030 a1074d23-0009-0410-80fe-cf8c14f379e6