Fix (pdfwrite) : Generalize pdf_finish_font_descriptors (continued).
DETAILS :

This is a preparation for the bug 688790
"pdfwrite : The PDF/A conversion needs to to re-encode texts".
The last patch appears incomplete.
This adds the pdf_resource_type_t parameter to pdf_finish_resources,
and defines pdf_convert_truetype_font_descriptor for future implementation.

This change is algorithmically equivalent.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@7054 a1074d23-0009-0410-80fe-cf8c14f379e6