Fix missing current point after painting path in PXL.
Graphics library after r4886 no longer uses the path to store the
current point rather it does so explicitly.  Some functions that used to
implicitly set the current point not longer do so requiring a pair of
functions to do so.

In this case copying the path didn't set the current point.

git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@2588 06663e23-700e-0410-b217-a244a6096597