Changes to ttf fonts to allow use of ufst without FAPI.
Truetype fonts can be processed either by FAPI, straight ufst or ttfReader
The decision is contolled in gx_add_fm_pair() consulting varibles in the font.
If FAPI is used this is signaled by setting the pfont->FAPI != NULL in the
base font. else a new flag (USE_ttfReader) in the gs_type42 font controls
if ttfReader is to be used or if the regular font functions are to be used
in the current case this is used for pcl/pxl ttf via ufst.

This could be concidered a new subclass of gs_type42 with the same font type,
so that other ISA checks of font type treat the font as a ttf.

Fixes: Downloaded ttf fonts from PXL using UFST.

git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@2589 06663e23-700e-0410-b217-a244a6096597