Have the configure script check for fopen64 before using it. Bug 688394.
DETAILS:

We use this as a shorthand to check for the marked *64() stdio
calls, and based on that define HAVE_FILE64, which is used in
the unix platform code to select whether to call the marked
or normal FILE code when _LARGEFILE64_SOURCE is not defined.

This is required on MacOS X which does not provide fopen64()
and friends (or O_LARGEFILE) but nevertheless implements
reasonable large file support with the normal routines on
32-bit architectures without defining _LARGEFILE*_SOURCE.

The result should be the same with the autoconf build on
architectures where the previous code worked, including
linux and cygwin. When building manually without passing
-DHAVE_FILE64, or on platforms that don't provide the
marked 64-bit versions, we will now fall back to the
vanilla functions, and so the bug may re-occur there. This
is better than failing to link on such platforms, which
was the previous behavior.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@7070 a1074d23-0009-0410-80fe-cf8c14f379e6