Add an error callback to the example jasper applications. Previously they always failed (or succeeded) silently.
DETAILS:

The fprintf() calls were replaced with jas_eprintf() in this code along
with everywhere else, but no actual callback was installed, so the
applications weren't able to give any feedback at all. So we now provide
a simple callback that just passes the message on to stderr as is appropriate
for a library client.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@7073 a1074d23-0009-0410-80fe-cf8c14f379e6