NOTICES:
1) It is recommended to do a clean build with this update.

2) This makes the (now notorious) "use ttfreader" patch irrelevant.  So
that code can be backed out.  For now pcl sets the second parameter of
gs_type42_font_init() to true.  When the patch is backed out the
second parameter to use the ttfreader should be removed from the pcl
code.

This patch brings available many of the fixes that were made in the
artfifex font scaler systems to ufst based systems.  It also refactors
much of the pcl shared font code to reduce duplication of code.

We will compile a list of fixes of Genoa files associated with this
change when time permits, we want to have this code checked in now so
other changes like (2) can go forward.

git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@2611 06663e23-700e-0410-b217-a244a6096597