Add a simple 'smoke test' script for checking the language switch build.
If you have the pcl test suite in a directory 'tests' on the same level as
the ghostpcl source tree, you can run 'make check' (from the ghostpcl directory)
to invoke the script. It will run a selection of files through the local
build of pspcl6 and compare checksums against a baseline included with
this commit.

On my machine, the check takes about a minute and the current revision
shows 25 rendering differences compared to a baseline of r2611+7094.

To update the baseline, run './tools/smoke_update.sh'.

Caveats: The script can only be invoked from a particular directory.
It uses the md5sum of the ppmraw output, so if GS_PRODUCT changes
the checksums will no longer match. Updating the baseline will probably
fail if not run from an svn working copy. There is no support of updating
the baseline for only particular files.

git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@2622 06663e23-700e-0410-b217-a244a6096597