Fix for 688880 Encrypted PDF can not be processed.
DETAILS:

The problem file specifies both V = 1 and and a Length = 128
bits in the encryption dict.  This is contradictory since page 93 of
the PDF 1.6 spec says that if V = 1 then the key length should be
40 bits.

This change ignores the the Length value if V = 1.  This matches the
observed behavior with Adobe Acrobat 7.0.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@7112 a1074d23-0009-0410-80fe-cf8c14f379e6