Fix (psi) : Provide a right finalization for the Postscript interpreter, part 2.
DETAILS :

This patch requires gs revision 7111.

1. The old code placed a device structure in Postscript's global memory,
and didn't inform Postscripot about that. While Postscript finalization
the device was freed because Postscxript memory contains no references to it.
This patch adds a code for registering the device
as a root so that the garbager doesn't free it. It also
unregisters the root when the device is not longer in use.

2. Provide a right inplementation for ps_impl_remove_device
(not important for the subject, but commonly useful).

3. Change pl_top_create_device from public to private
for a better reflection of assumptions done for this patch.

4. Insert gsapi_exit, which was removed in the revision 2643.
The log message of that revision appears confusing.
Actually the deacclosation code was correct,
and the allocation code was not - see (1).

5. (minor change) Exclude an unuseful type cast in psitop.c near 'dmem'.

Foe more details see comments in code.

Debugged with chars.pcl .

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@2648 06663e23-700e-0410-b217-a244a6096597