Fix : Specialize the build_char function argument type.
DETAILS :

The ghostpcl project yielded a potentially harmful
warning about a type cast. It occurs a cast from
a class to subclass. We could not provide a dynamic check in there,
so we analyzed why it needs such cast. The reason appears
an inaccurate design in the graphics library, and we fix it now.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@7127 a1074d23-0009-0410-80fe-cf8c14f379e6