reimplementation of passthrough to comply with the the PCL 4600 CET. Fixes Customer image #'s: 17, 18, 19 20, 21, 22, 23, 24, 25, 30 (improved), 37, 38, 39, 97.
Problems not passthrough related: improves #11.

This code will need review, cleanup, compiler warning purge, and more
documentation but it is functional.  There are also a few globals that
should be moved to the xl state when time permits.  The C705 file
which in itself is a complete test for pcl and pclxl needs to be
retested, minor discrepancies still exist.  The bugs noted above should
be closed and new problems opened.

git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@2669 06663e23-700e-0410-b217-a244a6096597