Fix (imagemask) : An optimization for imagemask with a shading color, continued.
DETAILS :

This is a performance fix for 09-47D-3 .
It is a similar optimization as a recent one for imagemask,
rather this one applies to different cases, which are not directly related to
to Postscript operator imagemask. The cases are :
- filling a cached character with a shading color;
- filling a FAPI character with a shading color;
- rendering an imagemask while a clist playback (for future.
  currently never happens with shadings).

This patch adds new modules gximask.c and gximask.h,
which factor out the common code for the mentioned cases.
The code is taken from the last patch for imagemask,
with minor restructurization.

The new function gx_image_fill_masked must be called
from the graphics librari instead a direct call
to the device virtual method fill_mask.
It checks for the shading color and inserts the clipping device,
which represents the mask.

Note that the clipping device always decompose a shading color,
so inserting the clipping device before it won't cause a problem.
However it needs some care when rendering to a high level device.

The clip accumulator device had no fill_mask virtual method.
This patch adds the default implementation
for the proper work of gx_image_fill_masked.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@7150 a1074d23-0009-0410-80fe-cf8c14f379e6