Turn on CPSI mode for CET runs. This mode uses the CPSI cmyk_to_rgb conversion and 'eofill' for Type1 characters.s
DETAILS:

This improves the difference count for 250 pages and only increases
the difference count for 6 pages. I've reviewed all regressions and
the changes are not important. Of the 250 pages that are improved,
The 250 improvements will be analyzed and those that were DIFF or
AOK will be changed to OK if they now match closely enough.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@7172 a1074d23-0009-0410-80fe-cf8c14f379e6