Fix (shadings) : Remove colors from C stack.
DETAILS :

This is the 4th step of fixing the bug 688955
"64K stack overflows with shadings".
This change is algorithmically equivalent.

Providing a color stack pointer field in the shading fill state structure.
When some colors are reserved, we advance the color stack pointer.
A new function release_colors work for releasing them.

This patch inserts some asserts, which to be removed soon.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@7190 a1074d23-0009-0410-80fe-cf8c14f379e6