Fix (shadings) : Radial shadings sometimes missed Extend.
DETAILS :

Debugged with CET 09-47I.PS SpecialTestH08Test06.

The old code wrongly computed parameter range for the extent,
when there exist 2 tangenting circles.
In this case the visible part of the extent prolongues to the farest circle.
The old code restricts the extent with the nearest circle.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@7218 a1074d23-0009-0410-80fe-cf8c14f379e6