Fix (stroking) : The CPSI compatibility about the strokepath starting point.
DETAILS :

Debugged with CET 11-21.PS page 10.

PLRM doesn't specify a starting point for strokepath result.
Now we choose same point as CPSI does.
It appears important if a further drawing depends on current point,
and/or when a dashing is applied to strokepath.

We don't keep the old algorithm because it is not useful.
Supporting two modes (the CPSI compatible mode and the GS native mode)
would be too complicated with no good reason.

EXPECTED DIFFERENCES :

"298-05.ps"
"321-05.ps"

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@7464 a1074d23-0009-0410-80fe-cf8c14f379e6