Fix : A CPSI compatibility for 'rectstroke'.
DETAILS :

Debugged with CET 12-12.PS page 8, 10.

PLRM doesn't specify the contour starting point for 'rectstroke'
when the rectangle width and/or height is negative.
CPSI and the old Ghostscript appear to produce different results,
which is important with dashing. The patch provides the CPSI
compatibility when CPSI_mode is true.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@7494 a1074d23-0009-0410-80fe-cf8c14f379e6