Fix : 'clippath' didn't set the current point.
DETAILS :

The bug has been introduced when we implemented the current point in 'double'.
It caused an inconsistent result when 'show' follows 'clippath',
because 'show' still uses pgs->ppath->position.
Debugged with CET 13-22.ps page 6, the item much13.

We would like to avoid the usage of pgs->ppath->position,
but currently we nave no time for a complete analyzis.
Delaying it for better times. See bug 689031.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@7531 a1074d23-0009-0410-80fe-cf8c14f379e6