Changes to the regression tools to facilitate parallel runs.
DETAILS:

Add a scratchdir entry to testing.cfg to specify where temporary files
should be created instead of using the current working directory.

Override various TextTestResult methods to print 'checking...result'
all at once instead of waiting while the job runs to reduce log
mangling. The old behaviour was helpful behavior in an interactive
serial run, but not running in batch mode.

If the PBS_NODENUM environment variable is set, use it to select a
custom testing.cfg.

If PBS_NODENUM is set, only run the source tests when it is zero to
avoid duplicating these.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@7575 a1074d23-0009-0410-80fe-cf8c14f379e6