Type 1 hinter : Implementing a contour sign normalization, part 2.
DETAILS :

This is continuation of the preparation for fixing the bug 688947
"Ghostscript renders font incorrectly (letter 'o' filled in)".

This change is algorithmically equivalent -
the new stuff works idle.
It define necesary data structures and flows
for storing information about subglyphs.
The implementation itself will be a next step.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@7583 a1074d23-0009-0410-80fe-cf8c14f379e6